requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
kagglehub>=0.2.0
pytest>=7.2.0
numpy
matplotlib 
seaborn 
scikit-learn 
pillow
